requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
sentence-transformers
python-dotenv
tavily-python
streamlit
faiss-cpu